SASS-5057: Review comments incorporated
