+++ app/uk/gov/hmrc/incometaxproperty/controllers/PeriodicSubmissionController.scala
@@ -20,14 +20,14 @@ import play.api.libs.json.Json
 import play.api.mvc.{Action, AnyContent, ControllerComponents}
 import uk.gov.hmrc.incometaxproperty.actions.AuthorisedAction
 import uk.gov.hmrc.incometaxproperty.models.errors.DataNotFoundError
-import uk.gov.hmrc.incometaxproperty.services.PropertyServices
+import uk.gov.hmrc.incometaxproperty.services.PropertyService
 import uk.gov.hmrc.play.bootstrap.backend.controller.BackendController
 
 import javax.inject.{Inject, Singleton}
 import scala.concurrent.ExecutionContext
 
 @Singleton()
-class PeriodicSubmissionController @Inject()(propertyServices: PropertyServices,
+class PeriodicSubmissionController @Inject()(propertyServices: PropertyService,
                                              authorisedAction: AuthorisedAction,
                                              cc: ControllerComponents)(implicit ec: ExecutionContext)
   extends BackendController(cc) {
