+++ app/uk/gov/hmrc/incometaxproperty/models/errors/ServiceError.scala
@@ -27,15 +27,3 @@ case object DataNotFoundError extends ServiceError {
   override val message: String = "User data could not be found due to mongo exception"
 }
 
-case object DataNotUpdatedError extends ServiceError {
-  override val message: String = "User data was not updated due to mongo exception"
-}
-
-case class MongoError(error: String) extends ServiceError {
-  override val message: String = s"Mongo exception occurred. Exception: $error"
-}
-
-case class EncryptionDecryptionError(error: String) extends ServiceError {
-  override val message: String = s"Encryption / Decryption exception occurred. Exception: $error"
-}
-
