+++ conf/application.conf
@@ -46,20 +46,6 @@ play.i18n.langs = ["en"]
 # !!!WARNING!!! DO NOT CHANGE THIS ROUTER
 play.http.router = prod.Routes
 
-# Controller
-# ~~~~~
-# By default all controllers will have authorisation, logging and
-# auditing (transaction monitoring) enabled.
-# The below controllers are the default exceptions to this rule.
-
-controllers {
-  uk.gov.hmrc.incometaxproperty.controllers.MicroserviceHelloWorldController = {
-    needsAuth = false
-    needsLogging = false
-    needsAuditing = false
-  }
-}
-
 # Microservice specific config
 
 mongodb {
