+++ it/uk/gov/hmrc/incometaxproperty/connectors/IntegrationFrameworkConnectorISpec.scala
@@ -21,9 +21,11 @@ import play.api.http.Status.{INTERNAL_SERVER_ERROR, OK}
 import play.api.libs.json.Json
 import uk.gov.hmrc.http.{HeaderCarrier, HttpResponse, SessionId}
 import uk.gov.hmrc.incometaxproperty.models.errors.{ApiError, SingleErrorBody}
+import uk.gov.hmrc.incometaxproperty.models.responses.{PeriodicSubmissionIdModel, PeriodicSubmissionModel}
 import uk.gov.hmrc.incometaxproperty.support.ConnectorIntegrationTest
 import uk.gov.hmrc.incometaxproperty.utils.builders.IncomeSourceDetailsBuilder.anIncomeSourceDetails
 
+import java.time.LocalDate
 import scala.concurrent.ExecutionContext.Implicits.global
 
 class IntegrationFrameworkConnectorISpec extends ConnectorIntegrationTest
@@ -52,6 +54,28 @@ class IntegrationFrameworkConnectorISpec extends ConnectorIntegrationTest
 
         await(underTest.getBusinessDetails(nino)(hc)) shouldBe
           Left(ApiError(INTERNAL_SERVER_ERROR, SingleErrorBody("some-code", "some-reason")))
+      }
+    }
+  }
+
+
+  val aPeriodicSubmissionModel = PeriodicSubmissionModel((List(
+    PeriodicSubmissionIdModel("1", "2021-01-01", "2021-11-11"),
+    PeriodicSubmissionIdModel("2", "2022-02-02", "2022-12-12")
+  )))
+
+  "Given a need to get Periodic Submission Data" when {
+    "a call is made to the backend API it" should {
+      "return correct IF data when correct parameters are passed" in {
+        val taxYear = LocalDate.now.getYear.toString
+        val taxableEntityId = "some-taxable-entity-id"
+        val incomeSourceId = "some-income-source-id"
+
+        val httpResponse = HttpResponse(OK, Json.toJson(aPeriodicSubmissionModel).toString())
+
+        stubGetHttpClientCall(s"/income-tax/business/property/$taxYear/$taxableEntityId/$incomeSourceId/period", httpResponse)
+
+        await(underTest.getPeriodicSubmission(taxYear, taxableEntityId, incomeSourceId)(hc)) shouldBe Right(aPeriodicSubmissionModel)
       }
     }
   }
