+++ project/plugins.sbt
@@ -6,7 +6,7 @@ ThisBuild / libraryDependencySchemes ++= Seq(
   "org.scala-lang.modules" %% "scala-xml" % VersionScheme.Always
 )
 
-addSbtPlugin("uk.gov.hmrc"       % "sbt-auto-build"         % "3.9.0")
+addSbtPlugin("uk.gov.hmrc"       % "sbt-auto-build"         % "3.15.0")
 addSbtPlugin("uk.gov.hmrc"       % "sbt-distributables"     % "2.2.0")
 addSbtPlugin("com.typesafe.play" % "sbt-plugin"             % "2.8.19")
 addSbtPlugin("org.scoverage"     % "sbt-scoverage"          % "2.0.6")
